Tile fixes and additions
More wood has been added, one of the planks was fixed, the others redrawn. The buns have been adapted to the new palette, and I ❤︎ Bunny has been added to the 3x2 bun to make it a 3x3 bun.

The MacOS Platinum decorations will be removed, so will Visy's original floor texture in a future update.