Fixes and new textures
Fixed the eyes on the buns, now you shouldn't be able to see the world through them.
Redone the stone wall texture ( temporary, maybe? It doesn't look particularly exciting but it isn't awful either. My eyes hurt from so much orange, though. )
Added a concrete wall texture, and an alternate chunk for it.
Added a metallic interior floor and wall with a downward light on it. Not too sure on how it looks, but i'm adding for testing.
Replaced Visy's placeholder tiles with more placeholder tiles. they look better on the editor than they do on the map though.

I may stick to make more garden-like textures, and maybe move to make pettable/collectible buns to take advantage of the billboard system.